Update menu item routine names for Carbon
Call CheckMenuItem() instead of CheckItem().  {Enable,Disable}MenuItem()
are forward- but not backward-compatible -- a call to EnableItem() could
be changed to EnableMenuItem() (if implemented) with no problem, but the
converse isn't true in all cases (and similarly for DisableItem()) -- so
preserve the calls to the old functions in the code, and forward them to
the newer functions when the older ones aren't available (as in Carbon).