Don't lock handles without cause
A common error in Mac application development is calling HLock() without
an actual need to do so.  For example:

  * merely accessing fields of a relocatable block
  * passing the handle to an OS or Toolbox function (e.g. CopyBits())
  * calling a function known not to "move memory" (e.g. BlockMove())[1]
  * dereferencing a handle that's always locked in the first place

GDHandles are always locked, so there's no need to call HLock().  Though
at least the developer called HSetState() and not HUnlock() (which along
with the error of relocking a GDHandle at all resulted in a nasty bug in
the Mercutio MDEF).  Not only are GDHandles forever locked, one scenario
merely reads the gdPMap field (and then dereferences the PixMapHandle to
read the pixelSize field without locking it, ironically).

[1] Obviously, BlockMove() moves memory in a literal sense.  However, it
doesn't cause relocatable Memory Manager blocks (a.k.a. handles) to move
or get purged.